Many Many Changes to MF code
I've done so many changes, I don't even know where to start with listing all of them.
So I'm just going to leave this commit message as is ... sorry future self :(
